+++ js/main.js
@@ -2,4 +2,7 @@ import {createOffers} from './data.js';
 import {insertOffer} from './popup.js';
 import {getRandomPositiveInteger} from './util.js';
 
-insertOffer(createOffers()[getRandomPositiveInteger(0,9)]);
+const offers = createOffers();
+const randomIndex = getRandomPositiveInteger(0,9);
+
+insertOffer(offers[randomIndex]);
